Update the minimum SDK constraint to 2.2.2 for kernel and front_end packages.
This allows us to use ui-as-code and set literal features in these
packages.

This should be safe because (a) when these packages are used within
the SDK itself, they are used with the most recent SDK (whose version
is greater than 2.2.2), and (b) the only significant dependency on
these packages in pub is from the analyzer, which already has a
minimum SDK version of 2.2.2.