Migration: rework handling of expressions of the form `a?.b += c`.
Previously we just had to store a single node representing the
nullability of a, since we knew we would visit `a?.b` just before
finishing up the handling of `a?.b += c`.  However, in a follow-up CL
we're going to have to change the visit order so that `a?.b` is
visited, then `c`, then `a?.b += c` can be finished up.  So we need a
map to store the association between expressions like `a?.b` and the
nullabilities of the corresponding `a`'s.