Migration: decouple DecoratedType from PotentialModification.
Previously we had a single class, DecoratedTypeAnnotation, which
coupled together DecoratedType and PotentialModification, allowing one
object to serve both roles.  But this won't work for typedef
instantiations, because they will need to be associated with a
potential modification, but they will be created using the shared
infrastructure for inferred types (which just creates a plain
DecoratedType).  The solution is to decouple the two roles into two
separate objects.