Make exitFunctionBody safer.
There's no functional change here; it was wrong to have mismatched
exit/enter of a function body before, and it's still wrong.  But this
change makes it easier to track down problems if an exception gets
thrown debug visiting, causing a mismatch; if that happens we can
temporarily comment out the assert to avoid covering up the exception.